+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * added storing hostname
 * added reading and filtering for hostnames
+* added reading visitors by weekday and hour
 * updated dependencies
 
 ## 6.14.2